feat: windows可通过文件打开多个窗口并重构主进程代码
